server/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=1.1.2",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.2.2",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.32.0",
]
[tool.ruff]
# Set the indentation width
indent-width = 8
